add `deploymentName` to `workerQueue.Create()`
- post rebase from main deployments can be selected via deploymentID or deploymentName
- address PR comments
  - refactor `worker-queue...` to `worker queue ...` in user facing output
  - refactor variable names to not have `WorkerQueue` as it is redundant
  - move workerqueue create command to `cmd/cloud/deployment_workerqueue.go`
  - refactor to use the `errors` package
  - remove the `--isDefault` flag as users can not create a default queue
  - improve message for queue names
