+++ cloud/deployment/workerqueue/workerqueue.go
@@ -1,86 +1,85 @@
 package workerqueue
 
 import (
+	"errors"
 	"fmt"
 	"io"
 
 	"github.com/astronomer/astro-cli/astro-client"
 	"github.com/astronomer/astro-cli/cloud/deployment"
-	"github.com/pkg/errors"
 )
 
 var (
-	ErrWorkerQueueDefaultOptions = errors.New("failed to get worker-queue default options")
-	ErrInvalidWorkerQueueOption  = errors.New("worker-queue option is invalid")
-	ErrCannotUpdateExistingQueue = errors.New("worker-queue exists")
+	ErrWorkerQueueDefaultOptions = errors.New("failed to get worker queue default options")
+	ErrInvalidWorkerQueueOption  = errors.New("worker queue option is invalid")
+	ErrCannotUpdateExistingQueue = errors.New("worker queue already exists")
 )
 
-func Create(ws, deploymentID, name string, isDefaultWQueue bool, wQueueMin, wQueueMax, wQueueConcurrency int, client astro.Client, out io.Writer) error {
+func Create(ws, deploymentID, deploymentName, name string, wQueueMin, wQueueMax, wQueueConcurrency int, client astro.Client, out io.Writer) error {
 	var (
-		requestedDeployment       astro.Deployment
-		err                       error
-		errHelp                   string
-		workerQueueToCreate       *astro.WorkerQueue
-		wQueueListToCreate        []astro.WorkerQueue
-		workerQueueDefaultOptions astro.WorkerQueueDefaultOptions
-		pool                      astro.NodePool
-		nodePoolIDForWorkerQueue  string
+		requestedDeployment astro.Deployment
+		err                 error
+		errHelp             string
+		queueToCreate       *astro.WorkerQueue
+		listToCreate        []astro.WorkerQueue
+		defaultOptions      astro.WorkerQueueDefaultOptions
+		pool                astro.NodePool
+		nodePoolID          string
 	)
 	// get or select the deployment
-	requestedDeployment, err = deployment.GetDeployment(ws, deploymentID, client)
+	requestedDeployment, err = deployment.GetDeployment(ws, deploymentID, deploymentName, client)
 	if err != nil {
 		return err
 	}
 
-	// TODO should we only get defaults once via init()
 	// get defaults for min-count, max-count and concurrency from API
-	workerQueueDefaultOptions, err = GetWorkerQueueDefaultOptions(client)
+	defaultOptions, err = GetWorkerQueueDefaultOptions(client)
 	if err != nil {
-		return err
+		return fmt.Errorf("%w: %s", ErrWorkerQueueDefaultOptions, err.Error())
 	}
 
 	// Get the default nodepool ID to use
 	for _, pool = range requestedDeployment.Cluster.NodePools {
 		if pool.IsDefault {
-			nodePoolIDForWorkerQueue = pool.ID
+			nodePoolID = pool.ID
+			break
 		}
 	}
 
 	// TODO user selects nodePoolID for creating a queue on non-default nodepools
 
-	workerQueueToCreate = &astro.WorkerQueue{
+	queueToCreate = &astro.WorkerQueue{
 		Name:       name,
-		IsDefault:  isDefaultWQueue,
-		NodePoolID: nodePoolIDForWorkerQueue,
+		IsDefault:  false, // cannot create a default queue
+		NodePoolID: nodePoolID,
 	}
-	workerQueueToCreate = setWorkerQueueValues(wQueueMin, wQueueMax, wQueueConcurrency, workerQueueToCreate, workerQueueDefaultOptions)
+	queueToCreate = setWorkerQueueValues(wQueueMin, wQueueMax, wQueueConcurrency, queueToCreate, defaultOptions)
 
-	err = IsWorkerQueueInputValid(workerQueueToCreate, workerQueueDefaultOptions)
+	err = IsWorkerQueueInputValid(queueToCreate, defaultOptions)
 	if err != nil {
 		return err
 	}
 
-	if QueueExists(requestedDeployment.WorkerQueues, workerQueueToCreate) {
+	if QueueExists(requestedDeployment.WorkerQueues, queueToCreate) {
 		// create does not allow updating existing queues
-		errHelp = fmt.Sprintf("use worker-queue update %s instead", workerQueueToCreate.Name)
-		return errors.Wrap(ErrCannotUpdateExistingQueue, errHelp)
+		errHelp = fmt.Sprintf("use worker queue update %s instead", queueToCreate.Name)
+		return fmt.Errorf("%w: %s", ErrCannotUpdateExistingQueue, errHelp)
 	}
 
-	// workerQueueToCreate does not exist so we add it
-	wQueueListToCreate = append(requestedDeployment.WorkerQueues, *workerQueueToCreate) //nolint
+	// queueToCreate does not exist so we add it
+	listToCreate = append(requestedDeployment.WorkerQueues, *queueToCreate) //nolint
 
-	// update the deployment with the new list of worker-queues
-	err = deployment.Update(requestedDeployment.ID, "", ws, "", 0, 0, 0, wQueueListToCreate, true, client)
+	// update the deployment with the new list of worker queues
+	err = deployment.Update(requestedDeployment.ID, "", ws, "", "", 0, 0, 0, listToCreate, true, client)
 	if err != nil {
 		return err
 	}
-	fmt.Fprintf(out, "worker-queue %s for %s in %s workspace created\n", workerQueueToCreate.Name, requestedDeployment.ID, ws)
+	fmt.Fprintf(out, "worker queue %s for %s in %s workspace created\n", queueToCreate.Name, requestedDeployment.ID, ws)
 	return nil
 }
 
 // setWorkerQueueValues sets values for MinWorkerCount, MaxWorkerCount and WorkerConcurrency
 // Default values are used if the user did not request any
-
 func setWorkerQueueValues(wQueueMin, wQueueMax, wQueueConcurrency int, workerQueueToCreate *astro.WorkerQueue, workerQueueDefaultOptions astro.WorkerQueueDefaultOptions) *astro.WorkerQueue {
 	if wQueueMin != 0 {
 		// use the value from the user input
@@ -109,7 +108,6 @@ func setWorkerQueueValues(wQueueMin, wQueueMax, wQueueConcurrency int, workerQue
 // GetWorkerQueueDefaultOptions calls the workerqueues query
 // It returns WorkerQueueDefaultOptions if the query succeeds
 // An error is returned if it fails
-
 func GetWorkerQueueDefaultOptions(client astro.Client) (astro.WorkerQueueDefaultOptions, error) {
 	var (
 		workerQueueDefaultOptions astro.WorkerQueueDefaultOptions
@@ -125,23 +123,22 @@ func GetWorkerQueueDefaultOptions(client astro.Client) (astro.WorkerQueueDefault
 // IsWorkerQueueInputValid checks if the requestedWorkerQueue adheres to the floor and ceiling set in the defaultOptions
 // if it adheres to them, it returns nil
 // ErrInvalidWorkerQueueOption is returned if min, max or concurrency are out of range
-
 func IsWorkerQueueInputValid(requestedWorkerQueue *astro.WorkerQueue, defaultOptions astro.WorkerQueueDefaultOptions) error {
 	var errorMessage string
 	if !(requestedWorkerQueue.MinWorkerCount >= defaultOptions.MinWorkerCount.Floor) ||
 		!(requestedWorkerQueue.MinWorkerCount <= defaultOptions.MinWorkerCount.Ceiling) {
 		errorMessage = fmt.Sprintf("min worker count must be between %d and %d", defaultOptions.MinWorkerCount.Floor, defaultOptions.MinWorkerCount.Ceiling)
-		return errors.Wrap(ErrInvalidWorkerQueueOption, errorMessage)
+		return fmt.Errorf("%w: %s", ErrInvalidWorkerQueueOption, errorMessage)
 	}
 	if !(requestedWorkerQueue.MaxWorkerCount >= defaultOptions.MaxWorkerCount.Floor) ||
 		!(requestedWorkerQueue.MaxWorkerCount <= defaultOptions.MaxWorkerCount.Ceiling) {
 		errorMessage = fmt.Sprintf("max worker count must be between %d and %d", defaultOptions.MaxWorkerCount.Floor, defaultOptions.MaxWorkerCount.Ceiling)
-		return errors.Wrap(ErrInvalidWorkerQueueOption, errorMessage)
+		return fmt.Errorf("%w: %s", ErrInvalidWorkerQueueOption, errorMessage)
 	}
 	if !(requestedWorkerQueue.WorkerConcurrency >= defaultOptions.WorkerConcurrency.Floor) ||
 		!(requestedWorkerQueue.WorkerConcurrency <= defaultOptions.WorkerConcurrency.Ceiling) {
 		errorMessage = fmt.Sprintf("worker concurrency must be between %d and %d", defaultOptions.WorkerConcurrency.Floor, defaultOptions.WorkerConcurrency.Ceiling)
-		return errors.Wrap(ErrInvalidWorkerQueueOption, errorMessage)
+		return fmt.Errorf("%w: %s", ErrInvalidWorkerQueueOption, errorMessage)
 	}
 	return nil
 }
@@ -149,7 +146,6 @@ func IsWorkerQueueInputValid(requestedWorkerQueue *astro.WorkerQueue, defaultOpt
 // QueueExists takes a []existingQueues and a queueToCreate as arguments
 // It returns true if queueToCreate exists in []existingQueues
 // It returns false if queueToCreate does not exist in []existingQueues
-
 func QueueExists(existingQueues []astro.WorkerQueue, queueToCreate *astro.WorkerQueue) bool {
 	for _, queue := range existingQueues {
 		if queue.ID == queueToCreate.ID {
